requirements.txt
# KimiClaw Master Coder - Requirements
streamlit>=1.28.0
pandas>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0
